handle out of range timestamp
Sometimes RTP timestamp get completely out of range for several packets.
This would lead to a pretty small/huge channel start/end timestamp,
which may make demuxer tending to insert lots of dummy packets and
eventually out of memory.

So there are two things we did to solve this problem:
1. Don't use max ts value in whole RTP flow, use the 'relative' last
   timestamp in the flow
2. Don't add out of range timestamp RTP packet into channel's packet
   buffer, replace it with dummy packets

To achive the first thing, we needs to compare RTP packet's seq num to
decide whether pkt1 should be arrived after than pkt2. So we impl
PartialOrd for SeqNum. The gist of the algorithm is to seed whether
pkt1's seq is in range (pkt2.seq, pkt2.seq + 32786], if so, pkt1 should
arrive after pkt2, otherwise it should arrive earlier.

The above algorithm is based on an assumption that we don't buffer too
much packets to perform the above algorithm. In ideal situation, RTP seq
num would be reused after 65535/50/60=21.84 minutes, it's quite long
actually. In the future we may needs to see how the algorithm works when
seq num jumping happens;